pebblo_safeloader/langchain/identity-rag-pgvector/requirements.txt
langchain==0.2.0
langchain-community==0.2.0 # For PebbloSafeLoader, PebbloRetrievalQA, VectorStore
langchain-google-community>=1.0.4 # For GoogleDriveLoader
langchain-openai>=0.1.7 # For OpenAI LLM and OpenAIEmbeddings
python-dotenv==1.0.0
tiktoken
requests==2.31.0
unstructured[all-docs]
psycopg2-binary
pgvector